+++ tsl/src/continuous_aggs/create.c
@@ -23,6 +23,7 @@
 #include <catalog/indexing.h>
 #include <catalog/pg_aggregate.h>
 #include <catalog/pg_collation.h>
+#include <catalog/pg_namespace.h>
 #include <catalog/pg_trigger.h>
 #include <catalog/pg_type.h>
 #include <catalog/toasting.h>
@@ -184,6 +185,7 @@ typedef struct CAggTimebucketInfo
 	Interval *interval;			  /* stores the interval, NULL if not specified */
 	const char *timezone;		  /* the name of the timezone, NULL if not specified */
 
+	FuncExpr *bucket_func; /* function call expr of the bucketing function */
 	/*
 	 * Custom origin value stored as UTC timestamp.
 	 * If not specified, stores infinity.
@@ -712,6 +714,24 @@ caggtimebucketinfo_init(CAggTimebucketInfo *src, int32 hypertable_id, Oid hypert
 	TIMESTAMP_NOBEGIN(src->origin); /* origin is not specified by default */
 }
 
+static Const *
+check_time_bucket_argument(Node *arg, char *position)
+{
+	if (IsA(arg, NamedArgExpr))
+		arg = (Node *) castNode(NamedArgExpr, arg)->arg;
+
+	Node *expr = eval_const_expressions(NULL, arg);
+
+	if (!IsA(expr, Const))
+		ereport(ERROR,
+				(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
+				 errmsg("only immutable expressions allowed in time bucket function"),
+				 errhint("Use an immutable expression as %s argument to the time bucket function.",
+						 position)));
+
+	return castNode(Const, expr);
+}
+
 /*
  * Check if the group-by clauses has exactly 1 time_bucket(.., <col>) where
  * <col> is the hypertable's partitioning column and other invariants. Then fill
@@ -722,6 +742,7 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 {
 	ListCell *l;
 	bool found = false;
+	bool custom_origin = false;
 
 	/* Make sure tbinfo was initialized. This assumption is used below. */
 	Assert(tbinfo->bucket_width == 0);
@@ -737,9 +758,16 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 			FuncExpr *fe = ((FuncExpr *) tle->expr);
 			Node *width_arg;
 			Node *col_arg;
-			Node *tz_arg;
 
 			if (!function_allowed_in_cagg_definition(fe->funcid))
+				continue;
+
+			/*
+			 * offset variants of time_bucket functions are not
+			 * supported at the moment.
+			 */
+			if (list_length(fe->args) >= 5 ||
+				(list_length(fe->args) == 4 && exprType(lfourth(fe->args)) == INTERVALOID))
 				continue;
 
 			if (found)
@@ -750,6 +778,8 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 			else
 				found = true;
 
+			tbinfo->bucket_func = fe;
+
 			/*only column allowed : time_bucket('1day', <column> ) */
 			col_arg = lsecond(fe->args);
 
@@ -759,54 +789,12 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 						 errmsg(
 							 "time bucket function must reference a hypertable dimension column")));
 
-			if (list_length(fe->args) == 4)
-			{
-				/*
-				 * Timezone and custom origin are specified. In this clause we
-				 * save only the timezone. Origin is processed in the following
-				 * clause.
-				 */
-				tz_arg = eval_const_expressions(NULL, lfourth(fe->args));
-
-				if (!IsA(tz_arg, Const))
-					ereport(ERROR,
-							(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
-							 errmsg("only immutable expressions allowed in time bucket function"),
-							 errhint("Use an immutable expression as fourth argument"
-									 " to the time bucket function.")));
-
-				Const *tz = castNode(Const, tz_arg);
-
-				/* This is assured by function_allowed_in_cagg_definition() above. */
-				Assert(tz->consttype == TEXTOID);
-				const char *tz_name = TextDatumGetCString(tz->constvalue);
-				if (!ts_is_valid_timezone_name(tz_name))
-				{
-					ereport(ERROR,
-							(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-							 errmsg("invalid timezone name \"%s\"", tz_name)));
-				}
-
-				tbinfo->timezone = tz_name;
-				tbinfo->bucket_width = BUCKET_WIDTH_VARIABLE;
-			}
-
 			if (list_length(fe->args) >= 3)
 			{
-				tz_arg = eval_const_expressions(NULL, lthird(fe->args));
-				if (!IsA(tz_arg, Const))
-					ereport(ERROR,
-							(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
-							 errmsg("only immutable expressions allowed in time bucket function"),
-							 errhint("Use an immutable expression as third argument"
-									 " to the time bucket function.")));
-
-				Const *tz = castNode(Const, tz_arg);
-				if ((tz->consttype == TEXTOID) && (list_length(fe->args) == 3))
+				Const *arg = check_time_bucket_argument(lthird(fe->args), "third");
+				if (exprType((Node *) arg) == TEXTOID)
 				{
-					/* Timezone specified */
-					const char *tz_name = TextDatumGetCString(tz->constvalue);
-
+					const char *tz_name = TextDatumGetCString(arg->constvalue);
 					if (!ts_is_valid_timezone_name(tz_name))
 					{
 						ereport(ERROR,
@@ -817,51 +805,69 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 					tbinfo->timezone = tz_name;
 					tbinfo->bucket_width = BUCKET_WIDTH_VARIABLE;
 				}
-				else
+			}
+
+			if (list_length(fe->args) >= 4)
+			{
+				Const *arg = check_time_bucket_argument(lfourth(fe->args), "fourth");
+				if (exprType((Node *) arg) == TEXTOID)
 				{
-					/*
-					 * Custom origin specified. This is always treated as
-					 * a variable-sized bucket case.
-					 */
+					const char *tz_name = TextDatumGetCString(arg->constvalue);
+					if (!ts_is_valid_timezone_name(tz_name))
+					{
+						ereport(ERROR,
+								(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
+								 errmsg("invalid timezone name \"%s\"", tz_name)));
+					}
+
+					tbinfo->timezone = tz_name;
 					tbinfo->bucket_width = BUCKET_WIDTH_VARIABLE;
-
-					if (tz->constisnull)
-					{
-						ereport(ERROR,
-								(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-								 errmsg("invalid origin value: null")));
-					}
-
-					switch (tz->consttype)
-					{
-						case DATEOID:
-							tbinfo->origin = DatumGetTimestamp(
-								DirectFunctionCall1(date_timestamp, tz->constvalue));
-							break;
-						case TIMESTAMPOID:
-							tbinfo->origin = DatumGetTimestamp(tz->constvalue);
-							break;
-						case TIMESTAMPTZOID:
-							tbinfo->origin = DatumGetTimestampTz(tz->constvalue);
-							break;
-						default:
-							/*
-							 * This shouldn't happen. But if somehow it does
-							 * make sure the execution will stop here even in
-							 * the Release build.
-							 */
-							ereport(ERROR,
-									(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
-									 errmsg("unsupported time bucket function")));
-					}
-
-					if (TIMESTAMP_NOT_FINITE(tbinfo->origin))
-					{
-						ereport(ERROR,
-								(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-								 errmsg("invalid origin value: infinity")));
-					}
 				}
+			}
+
+			/* check for custom origin */
+			switch (exprType(col_arg))
+			{
+				case DATEOID:
+					/* origin is always 3rd arg for date variants */
+					if (list_length(fe->args) == 3)
+					{
+						custom_origin = true;
+						tbinfo->origin = DatumGetTimestamp(
+							DirectFunctionCall1(date_timestamp,
+												castNode(Const, lthird(fe->args))->constvalue));
+					}
+					break;
+				case TIMESTAMPOID:
+					/* origin is always 3rd arg for timestamp variants */
+					if (list_length(fe->args) == 3)
+					{
+						custom_origin = true;
+						tbinfo->origin =
+							DatumGetTimestamp(castNode(Const, lthird(fe->args))->constvalue);
+					}
+					break;
+				case TIMESTAMPTZOID:
+					/* origin can be 3rd or 4th arg for timestamptz variants */
+					if (list_length(fe->args) >= 3 && exprType(lthird(fe->args)) == TIMESTAMPTZOID)
+					{
+						custom_origin = true;
+						tbinfo->origin =
+							DatumGetTimestampTz(castNode(Const, lthird(fe->args))->constvalue);
+					}
+					else if (list_length(fe->args) >= 4 &&
+							 exprType(lfourth(fe->args)) == TIMESTAMPTZOID)
+					{
+						custom_origin = true;
+						tbinfo->origin =
+							DatumGetTimestampTz(castNode(Const, lfourth(fe->args))->constvalue);
+					}
+			}
+			if (custom_origin && TIMESTAMP_NOT_FINITE(tbinfo->origin))
+			{
+				ereport(ERROR,
+						(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
+						 errmsg("invalid origin value: infinity")));
 			}
 
 			/*
@@ -895,35 +901,9 @@ caggtimebucket_validate(CAggTimebucketInfo *tbinfo, List *groupClause, List *tar
 						 errhint("Use an immutable expression as first argument"
 								 " to the time bucket function.")));
 
-			if ((tbinfo->bucket_width == BUCKET_WIDTH_VARIABLE) && (tbinfo->interval->month != 0))
+			if (tbinfo->interval && tbinfo->interval->month)
 			{
-				/* Monthly buckets case */
-				if (!TIMESTAMP_NOT_FINITE(tbinfo->origin))
-				{
-					/*
-					 * Origin was specified - make sure it's the first day of the month.
-					 * If a timezone was specified the check should be done in this timezone.
-					 */
-					Timestamp origin = tbinfo->origin;
-					if (tbinfo->timezone != NULL)
-					{
-						/* The code is equal to 'timestamptz AT TIME ZONE tzname'. */
-						origin = DatumGetTimestamp(
-							DirectFunctionCall2(timestamptz_zone,
-												CStringGetTextDatum(tbinfo->timezone),
-												TimestampTzGetDatum((TimestampTz) origin)));
-					}
-
-					const char *day =
-						TextDatumGetCString(DirectFunctionCall2(timestamp_to_char,
-																TimestampGetDatum(origin),
-																CStringGetTextDatum("DD")));
-					if (strcmp(day, "01") != 0)
-						ereport(ERROR,
-								(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-								 errmsg("for monthly buckets origin must be the first day of the "
-										"month")));
-				}
+				tbinfo->bucket_width = BUCKET_WIDTH_VARIABLE;
 			}
 		}
 	}
@@ -2312,18 +2292,17 @@ cagg_create(const CreateTableAsStmt *create_stmt, ViewStmt *stmt, Query *panquer
 		}
 
 		/*
-		 * `experimental` = true and `name` = "time_bucket_ng" are hardcoded
-		 * rather than extracted from the query. We happen to know that
-		 * monthly buckets can currently be created only with time_bucket_ng(),
-		 * thus these values are correct. Besides, they are not used for
+		 * These values are not used for
 		 * anything except Assert's yet for the same reasons. Once the design
 		 * of variable-sized buckets is finalized we will have a better idea
 		 * of what schema is needed exactly. Until then the choice was made
 		 * in favor of the most generic schema that can be optimized later.
 		 */
 		create_bucket_function_catalog_entry(materialize_hypertable_id,
-											 true,
-											 "time_bucket_ng",
+											 get_func_namespace(
+												 origquery_ht->bucket_func->funcid) !=
+												 PG_PUBLIC_NAMESPACE,
+											 get_func_name(origquery_ht->bucket_func->funcid),
 											 bucket_width,
 											 origin,
 											 origquery_ht->timezone);